pyproject.toml
[project]
name = "kind-words-for-techers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "passlib[bcrypt]==1.7.4",
    "pydantic==2.5.0",
    "pydantic-settings>=2.2.1",
    "python-jose[cryptography]==3.3.0",
    "python-multipart==0.0.6",
    "sqlalchemy==2.0.23",
    "uuid>=1.30",
    "uvicorn==0.24.0",
]
